requirements.txt
disnake==2.8.0
python-dotenv>=0.21.0